+++ ng/practices-ng-forms/src/lib/form-group-fluent-builder/extensions.ts
@@ -41,9 +41,7 @@ function toFormControlOptions<T>(
     controlDef: FormControlDefinition<T> | FormControlDefinitionValueOmitted
 ): FormControlOptions {
     return {
-        validators: [],
-        asyncValidators: [],
-        updateOn: controlDef.updateOn,
+        ...controlDef,
     };
 }